+++ packages/backend/gateway/src/controllers/test.controller.ts
@@ -12,90 +12,6 @@ export class TestEventsController {
     @Inject('EVENT_SERVICE') private readonly eventServiceClient: ClientProxy
   ) {}
 
-  @Get('events/1/activities')
-  @ApiOkResponse({
-    description: 'List of events'
-  })
-  public async getEventActivities(
-    @Res({ passthrough: true }) res: Response
-  ): Promise<any> {
-    res.header('x-total-count', String(3))
-    res.header('x-total-page', String(1))
-    return {
-      message: 'Lista de Atividades do Evento',
-      data: [
-        {
-          id: 1,
-          name: 'Competição Baiana de Veículos Autônomos em Escala',
-          activitieType: 'Competição',
-          workload: 10,
-          start_date: '22/09/1992',
-          end_date: '22/09/1992'
-        },
-        {
-          id: 2,
-          name: 'Competição Baiana de Veículos Autônomos em Escala',
-          activitieType: 'Palestra',
-          workload: 10,
-          start_date: '22/09/1992',
-          end_date: '22/09/1992'
-        },
-        {
-          id: 3,
-          name: 'Competição Baiana de Veículos Autônomos em Escala',
-          activitieType: 'Palestra',
-          workload: 10,
-          start_date: '22/09/1992',
-          end_date: '22/09/1992'
-        }
-      ],
-      errors: null
-    }
-  }
-
-  @Get('events')
-  @ApiOkResponse({
-    type: GetEventsResponseDto,
-    description: 'List of events'
-  })
-  public async getEvents(): Promise<GetEventsResponseDto> {
-    return {
-      message: 'Lista de Eventos',
-      data: {
-        events: [
-          {
-            name: 'Competição Baiana de Veículos Autônomos em Escala',
-            initials: 'CBVAE',
-            year: '2019',
-            start_date: new Date(),
-            end_date: new Date(),
-            description: 'Competição Baiana de Veículos Autônomos em Escala',
-            edition: ''
-          },
-          {
-            name: 'Projeto de Extensão do NAPNEE 	Curso LIBRAS',
-            initials: 'LIBRAS',
-            year: '2018',
-            start_date: new Date(),
-            end_date: new Date(),
-            description: 'Projeto de Extensão do NAPNEE 	Curso LIBRAS',
-            edition: ''
-          },
-          {
-            name: 'Projeto de Extensão do NAPNEE 	Curso TEA',
-            initials: 'TEA',
-            year: '2017',
-            start_date: new Date(),
-            end_date: new Date(),
-            description: 'Projeto de Extensão do NAPNEE 	Curso TEA',
-            edition: ''
-          }
-        ]
-      },
-      errors: null
-    }
-  }
-
   @Get('participants')
   @ApiOkResponse({
     type: GetEventsResponseDto,
