+++ packages/backend/gateway/src/interfaces/event/dto/activity/event-activity.dto.ts
@@ -1,0 +1,9 @@
+import { ApiProperty } from '@nestjs/swagger'
+
+export class ActivityEventIdDto {
+  @ApiProperty()
+  idEvent: string
+
+  @ApiProperty()
+  id: string
+}
