+++ packages/backend/gateway/src/interfaces/event/dto/event/event-participant.dto.ts
@@ -1,0 +1,9 @@
+import { ApiProperty } from '@nestjs/swagger'
+
+export class ParticipantEventIdDto {
+  @ApiProperty()
+  idEvent: string
+
+  @ApiProperty()
+  id: string
+}
