+++ packages/backend/gateway/src/interfaces/event/dto/get-events-response.dto.ts
@@ -7,30 +7,23 @@ export class GetEventsResponseDto {
   message: string
 
   @ApiProperty({
-    example: {
-      events: [
-        {
-          name: 'V Week-IT',
-          description:
-            'Cidades Inteligentes: Práticas Colaborativas no Sudoeste da Bahia',
-          initials: 'Week-IT',
-          year: '2019',
-          edition: '5º',
-          start_date: new Date('2019-10-01'),
-          end_date: new Date('2019-10-11'),
-          user_id: '5d987c3bfb881ec86b476bca',
-          created_at: new Date(),
-          updated_at: new Date(),
-          id: '5d987c3bfb881ec86b476bcc'
-        }
-      ]
-    },
+    example: [
+      {
+        name: 'V Week-IT',
+        description:
+          'Cidades Inteligentes: Práticas Colaborativas no Sudoeste da Bahia',
+        initials: 'Week-IT',
+        year: '2019',
+        edition: '5º',
+        start_date: new Date('2019-10-01'),
+        end_date: new Date('2019-10-11'),
+        user_id: '5d987c3bfb881ec86b476bca',
+        created_at: new Date(),
+        updated_at: new Date(),
+        id: '5d987c3bfb881ec86b476bcc'
+      }
+    ],
     nullable: true
   })
-  data: {
-    events: IEvent[]
-  }
-
-  @ApiProperty({ example: 'null' })
-  errors: { [key: string]: any }
+  data: IEvent[]
 }
