+++ packages/backend/gateway/src/interfaces/event/service-event-get-by-id-response.interface.ts
@@ -1,0 +1,9 @@
+import { IEvent } from './event.interface'
+
+export interface IServiceEventGetByIdResponse {
+  status: number
+  message: string
+  data: {
+    user: IEvent | null
+  }
+}
