+++ packages/backend/gateway/src/interfaces/event/service-event-list-response.interface.ts
@@ -1,0 +1,11 @@
+import { IEvent } from './event.interface'
+
+export interface IServiceEventListResponse {
+  status: number
+  message: string
+  data: {
+    events: IEvent[]
+    totalPages: number
+    totalCount: number
+  }
+}
