+++ packages/backend/gateway/src/interfaces/user/service-user-list-response.interface.ts
@@ -1,12 +1,11 @@
-import { IUser } from "./user.interface";
+import { IUser } from './user.interface'
 
 export interface IServiceUserListResponse {
-    status: number
-    message: string
-    data: {
-      users: IUser[]
-      totalPages: number
-      totalCount: number
-    }
+  status: number
+  message: string
+  data: {
+    users: IUser[]
+    totalPages: number
+    totalCount: number
   }
-  
+}
