+++ packages/backend/permission/src/constants/admin-permissions.ts
@@ -5,6 +5,11 @@ export const adminPermissions = [
   'user_confirm',
   'user_delete_by_id',
   'user_update_by_id',
+  'event_list',
+  'event_get_by_id',
+  'event_activity_list',
+  'event_activity_get_by_id',
+  'event_participant_list',
   'event_search_by_user_id',
   'event_create',
   'event_delete_by_id',
