+++ packages/frontend/web/src/styles/components/modal.ts
@@ -2,14 +2,20 @@ import { transparentize } from 'polished'
 import styled, { css } from 'styled-components'
 
 interface ModalProps {
-  size?: 'sm'
+  size?: 'sm' | 'lg'
   reverse?: boolean
+}
+
+interface RowProps {
+  cols: 2 | 3
 }
 
 export const Container = styled.div<ModalProps>`
   .modal-container-div {
     ${props =>
       props.size === 'sm' && 'width:' + props.theme.modal.size.sm + ';'}
+    ${props =>
+      props.size === 'lg' && 'width:' + props.theme.modal.size.lg + ';'}
   }
 
   display: none;
@@ -63,16 +69,20 @@ export const Container = styled.div<ModalProps>`
     border-top: 2px solid ${props => props.theme.colors.lightShade};
     padding: 20px 30px;
     display: flex;
-    justify-content: space-between;
+
+    ${props => props.size === 'sm' && 'justify-content: space-between;'}
+    ${props => props.size === 'lg' && 'justify-content: flex-end;'}
+ > button {
+      ${props => props.size === 'lg' && 'margin-left: 8px;'}
+    }
     ${props =>
       props.reverse &&
       css`
         flex-direction: row-reverse;
       `}
     > * {
-      flex: 1;
-      margin-left: 8px;
-      margin-right: 8px;
+      ${props =>
+        props.size === 'sm' && 'flex: 1; margin-left: 8px; margin-right: 8px;'}
       &:first-child {
         ${props => (props.reverse ? 'margin-right' : 'margin-left:') + '0;'}
       }
@@ -81,6 +91,13 @@ export const Container = styled.div<ModalProps>`
       }
     }
   }
+`
+
+export const Row = styled.div<RowProps>`
+  display: inline-grid;
+  ${props => props.cols === 2 && 'grid-template-columns: 1fr 1fr;'}
+  ${props => props.cols === 3 && 'grid-template-columns: 1fr 1fr 1fr;'}
+    grid-gap: 10px;
 `
 
 export const CloseArea = styled.div`
